feat: add share service
